break apart sudoku.py
